fix: fix methods calls after renamings
