+++ tests/test_crawler.py
@@ -19,8 +19,8 @@ from yarl import URL
 
 from udata_hydra import config
 from udata_hydra.analysis.resource import analyse_resource
-from udata_hydra.crawl import check_catalog
-from udata_hydra.crawl.check_resource import (
+from udata_hydra.crawl import start_checks
+from udata_hydra.crawl.check_resources import (
     RESOURCE_RESPONSE_STATUSES,
     check_resource,
 )
@@ -82,7 +82,7 @@ async def test_catalog_deleted_with_checked_resource(
 
     rurl = "https://example.com/resource-1"
     rmock.get(rurl)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
 
     res = await db.fetch("SELECT id FROM catalog WHERE deleted = FALSE and last_check IS NOT NULL")
     assert len(res) == 1
@@ -152,7 +152,7 @@ async def test_crawl(setup_catalog, rmock, event_loop, db, resource, analysis_mo
     # mock for head fallback
     rmock.get(rurl, **params)
     rmock.put(udata_url)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
 
     # test check results in DB
@@ -201,7 +201,7 @@ async def test_backoff_nb_req(setup_catalog, event_loop, rmock, mocker, fake_che
     mocker.patch("udata_hydra.config.BACKOFF_PERIOD", 0.25)
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # verify that we actually backed-off
     assert ("HEAD", URL(rurl)) not in rmock.requests
 
@@ -232,7 +232,7 @@ async def test_backoff_rate_limiting(
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=200)
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # verify that we actually backed-off
     if should_backoff:
         assert ("HEAD", URL(rurl)) not in rmock.requests
@@ -381,7 +381,7 @@ async def test_no_backoff_domains(
     mocker.patch("udata_hydra.context.monitor").return_value = magic
     rurl = "https://example.com/resource-1"
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # verify that we actually did not back-off
     assert not magic.add_backoff.called
 
@@ -391,7 +391,7 @@ async def test_excluded_clause(setup_catalog, mocker, event_loop, rmock, produce
     mocker.patch("udata_hydra.config.EXCLUDED_PATTERNS", ["http%example%"])
     rurl = "https://example.com/resource-1"
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # url has not been called due to excluded clause
     assert ("GET", URL(rurl)) not in rmock.requests
 
@@ -400,7 +400,7 @@ async def test_outdated_check(setup_catalog, rmock, fake_check, event_loop, prod
     await fake_check(created_at=datetime.now() - timedelta(weeks=52))
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # url has been called because check is outdated
     assert ("HEAD", URL(rurl)) in rmock.requests
 
@@ -412,7 +412,7 @@ async def test_not_outdated_check(
     await fake_check()
     rurl = "https://example.com/resource-1"
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # url has not been called because check is fresh
     assert ("GET", URL(rurl)) not in rmock.requests
 
@@ -421,7 +421,7 @@ async def test_switch_head_to_get(setup_catalog, event_loop, rmock, produce_mock
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=501)
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
     assert ("GET", URL(rurl)) in rmock.requests
 
@@ -430,7 +430,7 @@ async def test_switch_head_to_get_headers(setup_catalog, event_loop, rmock, prod
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=200, headers={})
     rmock.get(rurl, status=200)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
     assert ("GET", URL(rurl)) in rmock.requests
 
@@ -438,7 +438,7 @@ async def test_switch_head_to_get_headers(setup_catalog, event_loop, rmock, prod
 async def test_no_switch_head_to_get(setup_catalog, event_loop, rmock, produce_mock, analysis_mock):
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, status=200, headers={"content-length": "1"})
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
     assert ("GET", URL(rurl)) not in rmock.requests
 
@@ -505,7 +505,7 @@ async def test_analyse_resource_from_crawl(setup_catalog, rmock, event_loop, db,
     # mock for check and analysis results
     rmock.put(udata_url, status=200, repeat=True)
 
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
 
     assert len(rmock.requests[("PUT", URL(udata_url))]) == 2
     res = await db.fetch("SELECT * FROM checks")
@@ -521,7 +521,7 @@ async def test_change_analysis_last_modified_header(setup_catalog, rmock, event_
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     requests = rmock.requests[("PUT", URL(udata_url))]
     # last request is the one for analysis
     data = requests[-1].kwargs["json"]
@@ -539,7 +539,7 @@ async def test_change_analysis_content_length_header(
     rmock.head("https://example.com/resource-1", headers={"content-length": "2"})
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     requests = rmock.requests[("PUT", URL(udata_url))]
     # last request is the one for analysis
     data = requests[-1].kwargs["json"]
@@ -562,7 +562,7 @@ async def test_change_analysis_checksum(
     rmock.head("https://example.com/resource-1")
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     requests = rmock.requests[("PUT", URL(udata_url))]
     # last request is the one for analysis
     data = requests[-1].kwargs["json"]
@@ -583,7 +583,7 @@ async def test_change_analysis_harvested(
     mocker.patch("udata_hydra.analysis.resource.download_resource", mock_download_resource)
     rmock.head("https://example.com/harvested", headers={"content-length": "2"}, repeat=True)
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     requests = rmock.requests[("PUT", URL(udata_url))]
     # last request is the one for analysis
     data = requests[-1].kwargs["json"]
@@ -606,7 +606,7 @@ async def test_no_change_analysis_harvested(
     rmock.head("https://example.com/harvested", headers={"content-type": "application/json"})
     rmock.get("https://example.com/harvested")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("PUT", URL(udata_url)) not in rmock.requests
 
 
@@ -624,7 +624,7 @@ async def test_change_analysis_last_modified_header_twice(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has not been called: not first check, outdated check, and last-modified stayed the same
     assert ("PUT", URL(udata_url)) not in rmock.requests
 
@@ -645,7 +645,7 @@ async def test_change_analysis_last_modified_header_twice_tz(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has been called: last-modified has changed (different timezones)
     assert ("PUT", URL(udata_url)) in rmock.requests
     webhook = rmock.requests[("PUT", URL(udata_url))][0].kwargs["json"]
@@ -665,7 +665,7 @@ async def test_check_changed_content_length_header(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has been called in compute_check_has_changed: content-length has changed
     assert ("PUT", URL(udata_url)) in rmock.requests
     webhook = rmock.requests[("PUT", URL(udata_url))][0].kwargs["json"]
@@ -685,7 +685,7 @@ async def test_no_check_changed_content_length_header(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has not been called: not first check, outdated check, and content-length stayed the same
     assert ("PUT", URL(udata_url)) not in rmock.requests
 
@@ -703,7 +703,7 @@ async def test_check_changed_content_type_header(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has been called in compute_check_has_changed: content-type has changed
     assert ("PUT", URL(udata_url)) in rmock.requests
     webhook = rmock.requests[("PUT", URL(udata_url))][0].kwargs["json"]
@@ -723,7 +723,7 @@ async def test_no_check_changed_content_type_header(
     )
     rmock.get("https://example.com/resource-1")
     rmock.put(udata_url, repeat=True)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # udata has not been called: not first check, outdated check, and content-type stayed the same
     assert ("PUT", URL(udata_url)) not in rmock.requests
 
@@ -738,7 +738,7 @@ async def test_crawl_and_analysis_user_agent(setup_catalog, rmock, event_loop, p
     rurl = "https://example.com/resource-1"
     rmock.head(rurl, callback=callback)
     rmock.get(rurl, callback=callback)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
 
 
 async def test_check_triggered_by_udata_entrypoint_clean_catalog(
@@ -760,7 +760,7 @@ async def test_check_triggered_by_udata_entrypoint_clean_catalog(
     assert res.status == 200
     res = await db.fetch("SELECT * FROM catalog")
     assert len(res) == 1
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
     res = await db.fetch("SELECT * FROM checks")
     assert len(res) == 1
@@ -785,7 +785,7 @@ async def test_check_triggered_by_udata_entrypoint_existing_catalog(
     assert res.status == 200
     res = await db.fetch("SELECT * FROM catalog")
     assert len(res) == 2
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     assert ("HEAD", URL(rurl)) in rmock.requests
     res = await db.fetch("SELECT * FROM checks")
     assert len(res) == 2
@@ -803,7 +803,7 @@ async def test_check_triggers_csv_analysis(rmock, event_loop, db, produce_mock, 
         headers={"content-type": "application/csv"},
         body=SIMPLE_CSV_CONTENT.encode("utf-8"),
     )
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
     # GET called only once: HEAD is ok (no need for crawl) and analysis steps share the downloaded file
     assert len(rmock.requests[("GET", URL(rurl))]) == 1
     res = await db.fetch("SELECT * FROM checks")
@@ -831,7 +831,7 @@ async def test_recheck_download_only_once(
         },
     )
     await db.execute("UPDATE catalog SET priority = TRUE WHERE resource_id = $1", RESOURCE_ID)
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
 
     # HEAD should have been called
     assert len(rmock.requests[("HEAD", URL(rurl))]) == 1
@@ -862,7 +862,7 @@ async def test_dont_check_resources_with_status(
 ):
     await Resource.update(resource_id=RESOURCE_ID, data={"status": resource_status})
     rurl = "https://example.com/resource-1"
-    event_loop.run_until_complete(check_catalog(iterations=1))
+    event_loop.run_until_complete(start_checks(iterations=1))
 
     if resource_status == "BACKOFF" or resource_status is None:
         # HEAD should have been called
